# packages/types/openapi.yaml
openapi: 3.1.0
info:
  title: Mirage API
  version: 0.0.0
  description: |
    The single OpenAPI spec for Mirage. Lives in `@mirage/types` so it can
    be consumed both by the BFF (for route shape) and the React SPA (for the
    typed client). Generated TypeScript types land in `openapi.generated.ts`
    via `pnpm gen:openapi`, which is wired to `postinstall`.
servers:
  - url: http://localhost:4000
security:
  - bearerAuth: []
paths:
...
  /workspaces/{wsId}/sets/{id}/run:
    parameters:
      - in: path
        name: wsId
        required: true
        schema: { type: string }
      - in: path
        name: id
        required: true
        schema: { type: string }
    post:
      summary: Enqueue a full Run for a Set
      operationId: runSet
      responses:
        '201':
          description: Run created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Run' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
  /workspaces/{wsId}/runs:
    parameters:
      - in: path
        name: wsId
        required: true
        schema: { type: string }
    get:
      summary: List Runs in a workspace
      operationId: listRuns
      parameters:
        - in: query
          name: setId
          schema: { type: string }
        - in: query
          name: status
          schema: { type: string, enum: [queued, running, completed, failed, cancelled] }
        - in: query
          name: limit
          schema: { type: integer, minimum: 1, maximum: 500, default: 50 }
        - in: query
          name: offset
          schema: { type: integer, minimum: 0, default: 0 }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/RunListItem' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
  /workspaces/{wsId}/runs/{id}:
    parameters:
      - in: path
        name: wsId
        required: true
        schema: { type: string }
      - in: path
        name: id
        required: true
        schema: { type: string }
    get:
      summary: Get a single Run
      operationId: getRun
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Run' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
  /workspaces/{wsId}/runs/{id}/cancel:
    parameters:
      - in: path
        name: wsId
        required: true
        schema: { type: string }
      - in: path
        name: id
        required: true
        schema: { type: string }
    post:
      summary: Cancel a Run
      operationId: cancelRun
      responses:
        '204':
          description: Cancellation requested
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
  /workspaces/{wsId}/runs/{id}/preview:
    parameters:
      - in: path
        name: wsId
        required: true
        schema: { type: string }
      - in: path
        name: id
        required: true
        schema: { type: string }
    get:
      summary: Page through a Run's NDJSON artifact
      operationId: getRunPreview
      parameters:
        - in: query
          name: schemaKey
          required: true
          schema: { type: string }
        - in: query
          name: offset
          schema: { type: integer, minimum: 0, default: 0 }
        - in: query
          name: limit
          schema: { type: integer, minimum: 1, maximum: 1000, default: 200 }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/RunPreviewPage' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Run:
      type: object
      required: [id, orgId, workspaceId, setId, kind, status, requestedBy, createdAt]
      additionalProperties: false
      properties:
        id: { type: string }
        orgId: { type: string }
        workspaceId: { type: string }
        setId: { type: string }
        kind: { type: string, enum: [full, preview] }
        status: { type: string, enum: [queued, running, completed, failed, cancelled] }
        artifactKey: { type: string }
        rowCounts:
          type: object
          additionalProperties: { type: integer }
        startedAt: { type: string, format: date-time }
        endedAt: { type: string, format: date-time }
        errorMessage: { type: string }
        requestedBy: { type: string }
        createdAt: { type: string, format: date-time }
    RunListItem:
      type: object
      required: [id, setId, status, kind, requestedBy, createdAt]
      additionalProperties: false
      properties:
        id: { type: string }
        setId: { type: string }
        status: { type: string, enum: [queued, running, completed, failed, cancelled] }
        kind: { type: string, enum: [full, preview] }
        startedAt: { type: string, format: date-time }
        endedAt: { type: string, format: date-time }
        rowCounts:
          type: object
          additionalProperties: { type: integer }
        errorMessage: { type: string }
        requestedBy: { type: string }
        createdAt: { type: string, format: date-time }
    RunPreviewPage:
      type: object
      required: [schemaKey, offset, total, rows]
      additionalProperties: false
      properties:
        schemaKey: { type: string }
        offset: { type: integer, minimum: 0 }
        total: { type: integer }
        rows:
          type: array
          items: {}
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error: { type: string }
        code: { type: string }
        detail: {}
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
    Unauthorized:
      description: Unauthenticated
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
    Forbidden:
      description: Forbidden
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
    NotFound:
      description: Not found
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
...